The-AI-Engineer-Challenge/03_End-to-End_RAG_Clean/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
openai==1.3.0
python-dotenv==1.0.0
